+++ src/main/java/com/banghyang/chat/entity/Reason.java
@@ -1,0 +1,25 @@
+package com.banghyang.chat.entity;
+
+import com.banghyang.object.perfume.entity.Perfume;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class Reason {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String content;
+
+    @Builder
+    public Reason(String content) {
+        this.content = content;
+    }
+}
