+++ src/main/java/com/banghyang/chat/entity/Situation.java
@@ -1,0 +1,28 @@
+package com.banghyang.chat.entity;
+
+import com.banghyang.object.perfume.entity.Perfume;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class Situation {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String content;
+
+    @Builder
+    public Situation(String content) {
+        this.content = content;
+    }
+}
