+++ src/main/java/com/banghyang/chat/repository/ReasonRepository.java
@@ -1,0 +1,7 @@
+package com.banghyang.chat.repository;
+
+import com.banghyang.chat.entity.Reason;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReasonRepository extends JpaRepository<Reason, Long> {
+}
