+++ src/main/java/com/banghyang/chat/service/ImageGenerationService.java
@@ -1,8 +1,8 @@
-package com.banghyang.recommend.service;
+package com.banghyang.chat.service;
 
-import com.banghyang.recommend.entity.ChatImage;
-import com.banghyang.recommend.repository.ChatImageRepository;
+import com.banghyang.chat.repository.ChatImageRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
@@ -16,22 +16,17 @@ import java.util.Map;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class ImageGenerationService {
 
-    @Value("${base-url}")  // FastAPI 서버 URL 설정 (application.properties에서 관리)
+    @Value("${base-url}")  // FastAPI 서버 URL 설정 (application.yml 에서 관리)
     private String fastApiUrl;
 
     private final RestTemplate restTemplate;
     private final S3Service s3Service;
     private final ChatImageRepository chatImageRepository;
 
-    public ImageGenerationService(RestTemplate restTemplate, S3Service s3Service, ChatImageRepository chatImageRepository) {
-        this.restTemplate = restTemplate;
-        this.s3Service = s3Service;
-        this.chatImageRepository = chatImageRepository;
-    }
-
-    // FastAPI로 이미지를 전송하여 결과를 받음
+    // FastAPI 로 이미지를 전송하여 결과를 받음
     public Map<String, Object> generateImage(String prompt) {
         try {
             String url = fastApiUrl + "/image-generation/generate-image";
@@ -93,11 +88,12 @@ public class ImageGenerationService {
 
             String s3ImageUrl = s3Service.byteUploadImage(imageBytes, "generated-image.png");
 
-            ChatImage chatImage = ChatImage.builder()
-                    .imageUrl(s3ImageUrl)
-                    .build();
-            chatImageRepository.save(chatImage);
-            log.info("데이터베이스에 이미지 정보 저장 완료");
+            // 생성된 이미지는 url 만 필요하고 객체로써의 저장을 필요없다고 판단하여 삭제
+//            ChatImage chatImage = ChatImage.builder()
+//                    .imageUrl(s3ImageUrl)
+//                    .build();
+//            chatImageRepository.save(chatImage);
+//            log.info("데이터베이스에 이미지 정보 저장 완료");
 
             response.put("s3_url", s3ImageUrl);
             return response;
