+++ src/main/java/com/banghyang/chat/service/S3Service.java
@@ -1,4 +1,4 @@
-package com.banghyang.recommend.service;
+package com.banghyang.chat.service;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
