+++ src/main/java/com/banghyang/common/config/MongoConfig.java
@@ -1,4 +1,4 @@
-package com.banghyang.recommend.config;
+package com.banghyang.common.config;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
