+++ src/main/java/com/banghyang/common/config/RestTemplateConfig.java
@@ -1,4 +1,4 @@
-package com.banghyang.recommend.config;
+package com.banghyang.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
