+++ src/main/java/com/banghyang/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.banghyang.auth.common.config;
+package com.banghyang.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
