+++ src/main/java/com/banghyang/common/type/ChatType.java
@@ -1,0 +1,6 @@
+package com.banghyang.common.type;
+
+public enum ChatType {
+    USER,
+    AI
+}
