+++ src/main/java/com/banghyang/history/dto/historyResponse.java
@@ -1,0 +1,7 @@
+package com.banghyang.history.dto;
+
+import lombok.Data;
+
+@Data
+public class historyResponse {
+}
